+++ src/PolicyInsights/PolicyInsights/Properties/Resources.Designer.cs
@@ -19,7 +19,7 @@ namespace Microsoft.Azure.Commands.PolicyInsights.Properties {
     // class via a tool like ResGen or Visual Studio.
     // To add or remove a member, edit your .ResX file then rerun ResGen
     // with the /str option, or rebuild your VS project.
-    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "15.0.0.0")]
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "16.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     internal class Resources {
@@ -111,6 +111,15 @@ namespace Microsoft.Azure.Commands.PolicyInsights.Properties {
         internal static string Error_TooManyScopes {
             get {
                 return ResourceManager.GetString("Error_TooManyScopes", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized string similar to Triggering a policy compliance scan at scope &apos;{0}&apos;..
+        /// </summary>
+        internal static string StartComplianceScan {
+            get {
+                return ResourceManager.GetString("StartComplianceScan", resourceCulture);
             }
         }