+++ src/IotHub/IotHub/ChangeLog.md
@@ -22,6 +22,14 @@
     - `Get-AzIotHubDeviceTwin`
     - `Update-AzIotHubDeviceTwin`
 * Added cmdlet to invoke direct method on a device in an Iot Hub.
+* Manage IoT device module twin configuration, New cmdlets are:
+    - `Get-AzIotHubModuleTwin`
+    - `Update-AzIotHubModuleTwin`
+* Manage IoT automatic device management configuration at scale. New cmdlets are:
+    - `Add-AzIotHubConfiguration`
+    - `Get-AzIotHubConfiguration`
+    - `Remove-AzIotHubConfiguration`
+    - `Set-AzIotHubConfiguration`
 * Added cmdlet to invoke an edge module method in an Iot Hub.
 
 ## Version 2.3.0